Step-5 (in work -->  getCurrentController (2 variant-refactoring) )
