+++ controllers/authControllers/index.js
@@ -1,10 +1,12 @@
 const registrController = require("./registrController.js")
 const loginController = require("./loginController.js")
 const logoutController = require("./logoutController")
+const getCurrentController = require("./getCurrentController")
 
 
 module.exports = {
     registrController,
     loginController,
-    logoutController
+    logoutController,
+    getCurrentController
 }
