+++ controllers/index.js
@@ -1,10 +1,10 @@
 //! MAIN
 const contactsControllers = require("./contactsControllers");
 const authControllers = require("./authControllers")
-const usersControllers = require("./usersControllers")
+// const usersControllers = require("./usersControllers")
 
 module.exports = {
     contactsControllers,
     authControllers,
-    usersControllers
+    // usersControllers
 }
