+++ controllers/usersControllers/index.js
@@ -1,6 +1,6 @@
-const getCurrentController = require("./getCurrentController.js")
+// const getCurrentController = require("../authControllers/getCurrentController.js")
 
 
-module.exports = {
-    getCurrentController,
-}
+// module.exports = {
+//     getCurrentController,
+// }
