+++ routes/api/authRouter.js
@@ -24,11 +24,17 @@ router.post('/login', validateMiddlewarelogin, controllerWrapper(ctrl.loginContr
 
 
 //! 3-1. Проверка токена
-router.use(authMiddleware);
+router.use(authMiddleware); //! 1-вариант
 
 //! 3-2. Logout
 router.get('/logout', controllerWrapper(ctrl.logoutController)) //! 1-вариант
 // router.get('/logout', authMiddleware, controllerWrapper(ctrl.logoutController)) //! 2-вариант
+
+
+//! 4. Текущий пользователь - получить данные юзера по токену
+router.get('/current', controllerWrapper(ctrl.getCurrentController)) //! 1-вариант
+// router.get('/logout', authMiddleware, controllerWrapper(ctrl.getCurrentController)) //! 2-вариант
+
